Can send the stdout script in an email if an address is given as parameter
The test is a bit tricky because we need an SMTP server in order to check if the email is sent correctly.
To do that we instantiate a simple implementation in its own thread because the server is wainting in a loop
that would otherwise block the test.
Don't forget to stop the thread after the test.